Move SimplePie_Author to SimplePie\Author
