+++ library/SimplePie/Author.php
@@ -41,111 +41,15 @@
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
  */
 
-/**
- * Manages all author-related data
- *
- * Used by {@see SimplePie_Item::get_author()} and {@see SimplePie::get_authors()}
- *
- * This class can be overloaded with {@see SimplePie::set_author_class()}
- *
- * @package SimplePie
- * @subpackage API
- */
-class SimplePie_Author
-{
-	/**
-	 * Author's name
-	 *
-	 * @var string
-	 * @see get_name()
-	 */
-	var $name;
+ use SimplePie\Author;
 
-	/**
-	 * Author's link
-	 *
-	 * @var string
-	 * @see get_link()
-	 */
-	var $link;
+ class_exists('SimplePie\Author');
 
-	/**
-	 * Author's email address
-	 *
-	 * @var string
-	 * @see get_email()
-	 */
-	var $email;
+ // @trigger_error(sprintf('Using the "SimplePie_Author" class is deprecated since SimplePie version 1.x, use "SimplePie\Author" instead.'), \E_USER_DEPRECATED);
 
-	/**
-	 * Constructor, used to input the data
-	 *
-	 * @param string $name
-	 * @param string $link
-	 * @param string $email
-	 */
-	public function __construct($name = null, $link = null, $email = null)
-	{
-		$this->name = $name;
-		$this->link = $link;
-		$this->email = $email;
-	}
-
-	/**
-	 * String-ified version
-	 *
-	 * @return string
-	 */
-	public function __toString()
-	{
-		// There is no $this->data here
-		return md5(serialize($this));
-	}
-
-	/**
-	 * Author's name
-	 *
-	 * @return string|null
-	 */
-	public function get_name()
-	{
-		if ($this->name !== null)
-		{
-			return $this->name;
-		}
-
-		return null;
-	}
-
-	/**
-	 * Author's link
-	 *
-	 * @return string|null
-	 */
-	public function get_link()
-	{
-		if ($this->link !== null)
-		{
-			return $this->link;
-		}
-
-		return null;
-	}
-
-	/**
-	 * Author's email address
-	 *
-	 * @return string|null
-	 */
-	public function get_email()
-	{
-		if ($this->email !== null)
-		{
-			return $this->email;
-		}
-
-		return null;
-	}
-}
-
-class_alias('SimplePie_Author', 'SimplePie\Author', false);
+ if (\false) {
+     /** @deprecated since SimplePie 1.x, use "SimplePie\Author" instead */
+     class SimplePie_Author extends Author
+     {
+     }
+ }
